# contracts/openapi/guestbook.yaml
# HTTP contract for the guestbook. WRITTEN BY HAND -- do not dump openapi.json here.
#
# Compatibility mode: BACKWARD. A path, an operation, a response code and an
# optional field may be added. Removing a field, changing a type, narrowing a
# numeric bound and changing a stable refusal code are breaking and raise the
# major version.
#
# What stands here is frozen and nothing else: ./scripts/contracts.sh checks every
# sentence of this file against the Pydantic dump, and the dump is allowed to carry
# more. That is why there is no `operationId`, no `title`, and no `422` where it is
# only FastAPI's automatic one for a parameter -- nobody asked about them, so they
# cannot answer.
openapi: 3.0.3
info:
  title: Guestbook
  version: 1.3.0
  description: >-
    The rules behind these shapes are in spec/design/api.md and in
    spec/contexts/guestbook.md. This document states the field, the code and the
    cardinality in a form a machine can compare against the running application.

    Every `maxLength` and `minLength` below is in CODE POINTS, which is what JSON
    Schema means by the word and what this application now means by it on both
    sides of the wire. A value is normalized to NFC and trimmed of a written set of
    30 code points BEFORE it is measured, so a decomposed accented letter costs one
    and not two, and a signature of nothing but padding is empty rather than short.
    The set and the argument for it: app/platform/schemas/text.py.

paths:
  /api/guestbook-entries:
    get:
      summary: List entries (BR-04, BR-05)
      parameters:
        - name: q
          in: query
          required: false
        - name: sort
          in: query
          required: false
          # The one parameter whose set is CLOSED, so the one whose schema is named
          # rather than left to the application. The name is what reaches the browser:
          # `schema.d.ts` renders it as a union the frontend compiles against, and a
          # third word appearing on either side has to break something rather than
          # widen quietly. The other three parameters stay unfrozen here on purpose --
          # their bounds are numbers the register owns (spec/design/api.md
          # § Collection read parameters), not a set a caller can enumerate.
          schema:
            $ref: '#/components/schemas/GuestbookEntrySort'
        - name: limit
          in: query
          required: false
        - name: offset
          in: query
          required: false
      responses:
        "200":
          description: >-
            Always an envelope, never a bare array. An empty guestbook is a 200 with
            an empty `items`, never a 404: "there are no entries" is a successful
            answer to "list the entries".
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestbookEntryPage"
    post:
      summary: Add an entry (BR-01)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestbookEntryCreate"
      responses:
        "201":
          description: Exactly one entry.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestbookEntryRead"

  /api/guestbook-entries/{entry_id}:
    get:
      summary: Read one entry
      parameters:
        - name: entry_id
          in: path
          required: true
      responses:
        "200":
          description: Exactly one entry.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestbookEntryRead"
        "404":
          description: >-
            The identifier is valid, the entry is not there. The code
            `guestbook_entry_not_found` is in the body -- see x-refusals at the
            bottom of this file.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Refusal"
    patch:
      summary: Amend an entry (BR-02)
      parameters:
        - name: entry_id
          in: path
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestbookEntryUpdate"
      responses:
        "200":
          description: Exactly one entry, after the change.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestbookEntryRead"
        "404":
          description: The identifier is valid, the entry is not there.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Refusal"
        "422":
          description: >-
            Two different shapes under one code, and that is intended: Pydantic
            refuses a field that is empty or too long, and the router refuses a body
            that sets no field at all (`guestbook_entry_empty_patch`).
          # Frozen as a union from 1.2.0. Until then this entry carried a
          # `description` and no `content` -- and `scripts/openapi_contract.py`
          # skips any response with no JSON body, so it counted towards the
          # "responses checked" total while nothing about it was compared. The
          # router meanwhile declared `Refusal` alone, which is one of the two
          # shapes named just above. Both are named here now, so the document a
          # caller generates types from describes the response it will actually
          # get. The names are FastAPI's own and must stay: renaming a component
          # is a breaking change on its own.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: "#/components/schemas/Refusal"
                  - $ref: "#/components/schemas/HTTPValidationError"
    delete:
      summary: Delete an entry (BR-03)
      parameters:
        - name: entry_id
          in: path
          required: true
      responses:
        "204":
          description: Deleted. No body.
        "404":
          description: >-
            Also on a second deletion of the same entry -- deleting is not
            idempotent as far as the response is concerned (BR-03).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Refusal"

components:
  schemas:
    GuestbookEntryCreate:
      type: object
      required: [author, message]
      properties:
        # Code points after NFC and after trimming, on both sides of the wire.
        # The numbers did not move in 1.3.0; what moved is that the browser now
        # counts the same things the column and this document always counted.
        author:
          type: string
          minLength: 1
          maxLength: 80
        message:
          type: string
          minLength: 1
          maxLength: 1000

    GuestbookEntryUpdate:
      type: object
      properties:
        # No `type`: both fields are optional, so Pydantic renders them as an
        # `anyOf` with `null`. What is frozen is their EXISTENCE and the fact that
        # they are optional; writing `type: string` here would freeze a shape the
        # dump does not have.
        author: {}
        message: {}

    GuestbookEntryRead:
      type: object
      required: [id, author, message, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        author:
          type: string
        message:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    GuestbookEntryPage:
      type: object
      required: [items, total, total_all]
      properties:
        items:
          type: array
        total:
          type: integer
          minimum: 0
        total_all:
          type: integer
          minimum: 0

    Refusal:
      type: object
      required: [detail]
      properties:
        detail: {}

    RefusalDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        # No `type`: optional, so Pydantic renders an `anyOf` with `null`. What is
        # frozen is the field's existence and the fact that it is NOT required -- a
        # refusal about the whole request (an empty PATCH) has nothing to put in it.
        id: {}

    GuestbookEntrySort:
      type: string
      enum: [newest, oldest]

    # FastAPI's own validation shapes, frozen only because the 422 above now
    # refers to them: a `$ref` to a schema this document does not define is a
    # document that cannot be read on its own. Written to the same rule as
    # `Refusal` below -- what is stated is frozen and nothing else -- so `detail`
    # and `loc` freeze their existence and not their shape.
    HTTPValidationError:
      type: object
      properties:
        detail: {}

    ValidationError:
      type: object
      required: [loc, msg, type]
      properties:
        loc: {}
        msg:
          type: string
        type:
          type: string

# Stable refusal codes. They appear in NO OpenAPI document, because FastAPI does
# not look inside the body of a raised HTTPException -- and spec/design/api.md calls
# them a contract that "never changes". This list is the only place that holds them:
# ./scripts/contracts.sh looks for each one as a literal in every routers/ directory under app/.
x-refusals:
  guestbook_entry_not_found:
    status: 404
    when: the identifier is valid, the entry is not there -- also on a second deletion
  guestbook_entry_empty_patch:
    status: 422
    when: the PATCH sets no field at all